Add let keyword
Variable and function definitions start with the "let" keyword.

    let x = 7
    let y = 9

    let add = (int a, int b) {
        return a + b
    }

The exact function syntax hasn't been decided yet, but it will probably
be pretty close to that.

    let add = (int a, int b) {
        a + b
    }

and

    let add = (int a, int b) {
        return a + b;
    }

may be used.

We might want to move the `=` after the arguments, like an actual
function.

    let add (int a, int b) = {
        return a + b
    }

which would be equal to

    f(x, y) = x + y

but we may just want to throw out the equal and stick to C

    let add (int a, int b) {
        return a + b
    }

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>