Add AST and flesh out Parser
The parser now parses function and variable declarations and stores
them in an AST.

The syntax has changed. `let` is no longer used, but types can still be
infered. These are the same:

    int x = 7
    x = 7

Both of these set `x` equal to 7. The default number type will be a
or a Rat. Probably only a Rat when there's a decimal point. Or maybe
floats, depends on what's smaller.

Function syntax has reverted to C, with the inclusion of type
inference.

    int add(int a, int b) {
        return a + b
    }

is

    add(a, b) {
        return a + b
    }

though I don't know _how_ I'm going to do that. OCaml uses `+` for ints
and `+.` for floats, but that might get verbose.

The return value syntax isn't final. None of this is.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>