Change some types to Records and print Variables and Functions
Some of the types have been changed to Records, which fit better than a
variant type with one variation.

Variables and Functions are now printed to stdout. Variables are
printed with their type and name, and are set equal to the expression
they are assigned. Functions are stubs for now.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>