Create new Records for analyzing declarations
I ran out of tylenol trying to finish this function.

First, we match the type of declaration.
If it's a variable:
  Check if the type empty,
  then construct a new AST Record with the new type of the variable.
If it's a function:
  Go through every statement and check if it's a vairable declaration,
  then construct a new AST Record with the declaration of the variable
  set to the new type.

Each AST reconstruction is over 10 lines because "immutability is
good."

Changing a value in a Record, aka _making a new one_, has a simple
syntax:

  {record with value = new_value}

How about modifiying a List in a Record?

  {record with list = (List.map (fun x -> x + 1) record.list)}

That's a lot of boilerplate before we get to the actual checking part,
which includes a match statement with two paths.

  {record with list = (List.map (fun x ->
    match x with
    | TypeA a -> TypeA (a + 1)
    | TypeB b -> TypeB (b + 2)
  ) record.list)}

That's essentially what is going on. It gets a little more complicated
with the function statements.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>